Moves the Tanach rendering component into its own sub component
